Moving origin names to appsettings
