+++ ApplensBackend/appsettings.PublicAzure.json
@@ -4,5 +4,6 @@
     "certSubjectName": "applenstoruntimehost.appservice.compute.ce.azure.net",
     "endpoint": "https://wawsobserver.trafficmanager.net/api/",
     "diagnosticObserverEnabled": true
-  }
+  },
+  "IFrameAllowedOrigins": "https://azuresupportcenter.msftcloudes.com/ https://azuresupportcenterppe.msftcloudes.com/ https://azuresupportcentertest.azurewebsites.net/ https://eu.azuresupportcenter.azure.com"
 }
