+++ ApplensBackend/appsettings.USNat.json
@@ -77,5 +77,6 @@
     "Realm": "",
     "MetadataAddress": "",
     "Issuer": ""
-  }
+  },
+  "IFrameAllowedOrigins": "https://ex.azuresupportcenter.eaglex.ic.gov/"
 }
