+++ ApplensBackend/appsettings.USSec.json
@@ -77,5 +77,6 @@
     "Realm": "",
     "MetadataAddress": "",
     "Issuer": ""
-  }
+  },
+  "IFrameAllowedOrigins": "https://azuresupportcenter.microsoft.scloud/"
 }
