+++ ApplensBackend/appsettings.json
@@ -159,6 +159,7 @@
     "ResourceIDString": "AppServiceDiagnostics",
     "DocumentationStagingBranch": "DocumentationStagingBranch",
     "DocumentationRoot": "AppLensDocumentation",
-    "DiagnosticsMainBranch":  "master"
-  }
+    "DiagnosticsMainBranch": "master"
+  },
+  "IFrameAllowedOrigins":  ""
 }
